Jiminy's Journal Update
This consists of a rather robust feature update to include Jiminy's Journal alongside refactoring and a cleanup of the codebase. Many changes made to existing files were to clean up potential errors and warnings that were potentially slowing down build times. Added two new API calls: GetCharactersFirstAppearance and GetWorldsFirstAppearance. Also added was the refactoring of the JJCharacter into JournalEntry. This allowed for Characters, Enemies, Story Chronicles and Reports to be combined into one table and be sorted by Category. Also added was PJ's illustrations for MA, JJ and MS. The main things to make Jiminy's Journal functional is to setup the DailyJob to incorporate a random entry a day as well as the rest of the data.